Run tests concurrently.
This was done partly as a chance to "play" with C++11 threads.

A thread is created for each test case, which may be suboptimal. (Perhaps std::async could be used, which may use a set of worker threads - although the implementation on my PC may only use one thread.)

test completes quicker, however, this is currently offset by the increase in time to build it.